feat(example): add optional GPU support to Python example
Using "--use_gpu" as an argument in ./examples/gprat_python/ now
performs the example on the GPU. Otherwise and if GPRat was compiled
without CUDA, it executes the CPU example as done previously.

Additionaly, the output.csv file generated with the test, has an
additional Target column which contains "cpu" or "gpu" depending on the
Target used for computations.