feat(example): add optional GPU support to C++ example
Using "gpu --use_gpu" as an argument in ./examples/gprat_cpp/ now
performs the example on the GPU. Otherwise and if GPRat was compiled
without CUDA, it executes the CPU example as done previously. Using "gpu
cpu" uses the GPU Spack environment and uses the CPU for computations.
No additional argument or just "cpu" also performs computations on the
CPU using the CPU Spack environment. That is, it also compiles the
example script using the given Spack environment.

Additionaly, the output.csv file generated with the test, has an
additional Target column which contains "cpu" or "gpu" depending on the
Target used for computations. When performing the example on the GPU,
the GP optimize function is not executed and puts -1 into the respective
column.