feat(build): add dev-/release-linux-gpu to CMake presets
Added two CMake Presets dev-linux-gpu and release-linux-gpu to
CMakePresets.json. They compile GPRat with GPU support. In comparison,
the presets dev-linux and release-linux compile GPRat with CPU support.
Therefore, CI, which does not support GPUs currently, continues to work
as is. Adding the presets removes the necessity for cpu or gpu as the
second argument of compile_gprat.sh.

Also added appropriate documentation in README.md and fixed an error in
it.